medi-os/services/manage-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0
joblib>=1.3.0
scikit-optimize>=0.9.0
lightgbm>=4.0.0
catboost>=1.2.0
sentence-transformers>=2.2.0
pytorch-tabnet>=4.0
torch>=2.0.0
polars==1.12.0
google-generativeai>=0.8.0

